Split data/stack.s (#619)
* Split data/stack.s

* Separate files for audio start/end markers

* Split synthesis parts

* Split audio globals further into data, load and memory

* sSampleDmaNumListItems -> gSampleDmaNumListItems, and declare it in .h file

+++ src/audio/external.h
@@ -15,7 +15,7 @@
 extern s32 gAudioErrorFlags;
 extern f32 gDefaultSoundArgs[3];
 
-extern u64 gAudioSPTaskYieldBuffer[]; // ucode yield data ptr; only used in JP
+extern u8 gAudioSPTaskYieldBuffer[]; // ucode yield data ptr; only used in JP
 
 struct SPTask *create_next_audio_frame_task(void);
 void play_sound(s32 soundBits, f32 *pos);
